requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.1
Django==4.1.7
django-rest-knox==4.2.0
djangorestframework==3.14.0
psycopg2==2.9.5
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
